feat: rename add command
