+++ packages/cli/src/commands/add/eslint/eslint.ts
