+++ packages/cli/src/commands/add/eslint/index.ts
