+++ packages/cli/src/commands/add/eslint/templates/eslintrc-js.ts
