+++ packages/cli/src/commands/add/hydrogen/hydrogen.ts
