+++ packages/cli/src/commands/add/hydrogen/index.ts
