+++ packages/cli/src/commands/add/shopifyConfig/index.ts
