+++ packages/cli/src/commands/add/shopifyConfig/templates/shopify-config-js.ts
