+++ packages/cli/src/commands/add/shopifyProvider/templates/shopify-config-js.ts
