+++ packages/cli/src/commands/check/rules/eslint/checkEslintConfig.ts
@@ -1,5 +1,5 @@
 import {Env, CheckResult} from '../../../../types';
-import addEslint from '../../../addons/eslint';
+import addEslint from '../../../add/eslint';
 
 export async function checkEslintConfig({
   workspace,
